requirements.txt
solace-pubsubplus
python-dotenv
nuclia
supabase
